Add logic for settings page, modify constants, fix typos

+++ Pomodorek/Converters/TimeConverter.cs
@@ -8,8 +8,8 @@ public class TimeConverter : IValueConverter
 
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
-        var seconds = (short)value;
-        var minutes = (short)(seconds / 60);
+        var seconds = (int)value;
+        var minutes = seconds / 60;
         return $"{(minutes < 10 ? "0" : "")}{minutes}:{((seconds % 60) < 10 ? "0" : "")}{seconds % 60}";
     }
